// streampark-console/streampark-console-webapp-v2/src/enums/flinkEnum.ts
export enum DeployMode {
    /** standalone */
    STANDALONE = 1,
    /** yarn per-job (deprecated, please use yarn-application mode) */
    YARN_PER_JOB = 2,
    /** yarn session */
    YARN_SESSION = 3,
    /** yarn application */
    YARN_APPLICATION = 4,
    /** kubernetes session */
    KUBERNETES_SESSION = 5,
    /** kubernetes application */
    KUBERNETES_APPLICATION = 6,
}